Finish testing facade

+++ src/com/AlbertAbuav/facades/AdminFacade.java
@@ -5,8 +5,6 @@ import com.AlbertAbuav.beans.Coupon;
 import com.AlbertAbuav.beans.Customer;
 import com.AlbertAbuav.beans.CustomersVsCoupons;
 import com.AlbertAbuav.exceptions.invalidAdminException;
-import com.AlbertAbuav.exceptions.invalidCompanyException;
-import com.AlbertAbuav.exceptions.invalidCustomerException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -36,11 +34,14 @@ public class AdminFacade extends ClientFacade {
      * @param company Company
      */
     public void addCompany(Company company) throws invalidAdminException {
-        List<Company> companyList = companiesDAO.getAllCompanies();
-        if (companyList.size() == 0) {
+        if (company == null) {
+            throw new invalidAdminException("There is no customer like you entered!");
+        }
+        if (companiesDAO.getAllCompanies() == null) {
             companiesDAO.addCompany(company);
             return;
         }
+        List<Company> companyList = companiesDAO.getAllCompanies();
         for (Company c : companyList) {
             if (company.getName().equals(c.getName())) {
                 throw new invalidAdminException("The name of the company you are trying to add already appears in the system.\nCompanies with the same name cannot be added.");
@@ -59,6 +60,12 @@ public class AdminFacade extends ClientFacade {
      * @param company Company
      */
     public void updateCompany(Company company) throws invalidAdminException {
+        if (companiesDAO.getAllCompanies() == null) {
+            throw new invalidAdminException("There are no companies in the system!");
+        }
+        if (company == null) {
+            throw new invalidAdminException("There is no customer like you entered!");
+        }
         List<Company> companyList = companiesDAO.getAllCompanies();
         Company toCompare = null;
         for (Company company1 : companyList) {
@@ -81,21 +88,24 @@ public class AdminFacade extends ClientFacade {
      *
      * @param company Company
      */
-    public void deleteCompany(Company company) throws invalidCustomerException {
+    public void deleteCompany(Company company) throws invalidAdminException {
+        if (company == null) {
+            throw new invalidAdminException("There is no company like you entered!");
+        }
         if (companiesDAO.isCompanyExists(company.getEmail(), company.getPassword())) {
-            List<Coupon> companyCoupons = company.getCoupons();
-            if (companyCoupons.size() != 0) {
+            if (company.getCoupons() != null) {
+                List<Coupon> companyCoupons = company.getCoupons();
                 for (Coupon coupon : companyCoupons) {
                     List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCouponsByCouponId(coupon.getId());
                     for (CustomersVsCoupons purchase : purchases) {
-                        customersVsCouponsDAO.deleteCustomersVsCoupons(purchase.getCustomerID(), purchase.getCouponID());
+                        couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
                     }
                     couponsDAO.deleteCoupon(coupon.getId());
                 }
             }
             companiesDAO.deleteCompany(company.getId());
         } else {
-            throw new invalidCustomerException("There is no company by the name \"" + company.getName() + "\" in the system!");
+            throw new invalidAdminException("There is no company by the name \"" + company.getName() + "\" in the system!");
         }
     }
 
@@ -105,11 +115,10 @@ public class AdminFacade extends ClientFacade {
      * @return List
      */
     public List<Company> getAllCompanies() throws invalidAdminException {
-        List<Company> companies = companiesDAO.getAllCompanies();
-        if (companies.size() == 0) {
+        if (companiesDAO.getAllCompanies() == null) {
             throw new invalidAdminException("There are no companies in the system");
         }
-        return companies;
+        return companiesDAO.getAllCompanies();
     }
 
     /**
@@ -118,11 +127,14 @@ public class AdminFacade extends ClientFacade {
      * @param id int
      * @return Company
      */
-    public Company getSingleCompany(int id) {
+    public Company getSingleCompany(int id) throws invalidAdminException {
+        if (companiesDAO.getSingleCompany(id) == null) {
+            throw new invalidAdminException("No Company was found by this id!");
+        }
         Company company = companiesDAO.getSingleCompany(id);
-        List<Coupon> dbCoupons = couponsDAO.getAllCoupons();
-        List<Coupon> companyCoupons = new ArrayList<>();
-        if (dbCoupons.size() != 0) {
+        if (couponsDAO.getAllCoupons() != null) {
+            List<Coupon> dbCoupons = couponsDAO.getAllCoupons();
+            List<Coupon> companyCoupons = new ArrayList<>();
             for (Coupon dbCoupon : dbCoupons) {
                 if (dbCoupon.getCompanyID() == id) {
                     companyCoupons.add(dbCoupon);
@@ -141,18 +153,20 @@ public class AdminFacade extends ClientFacade {
      * @param customer Customer
      */
     public void addCustomer(Customer customer) throws invalidAdminException {
-        List<Customer> customerList = customersDAO.getAllCustomers();
-        if (customerList.size() == 0) {
+        if (customer == null) {
+            throw new invalidAdminException("There is no customer like you entered");
+        }
+        if (customersDAO.getAllCustomers() == null) {
             customersDAO.addCustomer(customer);
             return;
         }
+        List<Customer> customerList = customersDAO.getAllCustomers();
         for (Customer customer1 : customerList) {
             if (customer.getEmail().equals(customer1.getEmail())) {
                 throw new invalidAdminException("The email of the customer you are trying to add already appears in the system.\nCustomers with the same email cannot be added.");
             }
         }
         customersDAO.addCustomer(customer);
-
     }
 
     /**
@@ -162,6 +176,12 @@ public class AdminFacade extends ClientFacade {
      * @param customer Customer
      */
     public void updateCustomer(Customer customer) throws invalidAdminException {
+        if (customersDAO.getAllCustomers() == null) {
+            throw new invalidAdminException("There are no customers in the system!");
+        }
+        if (customer == null) {
+            throw new invalidAdminException("There is no customer like you entered");
+        }
         List<Customer> customerList = customersDAO.getAllCustomers();
         Customer toCompare = null;
         for (Customer customer1 : customerList) {
@@ -188,17 +208,20 @@ public class AdminFacade extends ClientFacade {
      *
      * @param customer Customer
      */
-    public void deleteCustomer(Customer customer) throws invalidCustomerException {
+    public void deleteCustomer(Customer customer) throws invalidAdminException {
+        if (customer == null) {
+            throw new invalidAdminException("This customer doesn't exists!");
+        }
         if (customersDAO.isCustomerExists(customer.getEmail(), customer.getPassword())) {
-            List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(customer.getId());
-            if (purchases.size() != 0) {
+            if (couponsDAO.getAllCustomersCoupons(customer.getId()) != null) {
+                List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(customer.getId());
                 for (CustomersVsCoupons purchase : purchases) {
-                    customersVsCouponsDAO.deleteCustomersVsCoupons(purchase.getCustomerID(), purchase.getCouponID());
+                    couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
                 }
+                customersDAO.deleteCustomer(customer.getId());
             }
-            customersDAO.deleteCustomer(customer.getId());
         } else {
-            throw new invalidCustomerException("There is no customer by the name \"" + customer.getFirstName() + " " + customer.getLastName() + "\" in the system!");
+            throw new invalidAdminException("There is no customer by the name \"" + customer.getFirstName() + " " + customer.getLastName() + "\" in the system!");
         }
     }
 
@@ -208,11 +231,10 @@ public class AdminFacade extends ClientFacade {
      * @return List
      */
     public List<Customer> getAllCustomers() throws invalidAdminException {
-        List<Customer> customers = customersDAO.getAllCustomers();
-        if (customers.size() == 0) {
+        if (customersDAO.getAllCustomers() == null) {
             throw new invalidAdminException("There are no customers in the system");
         }
-        return customers;
+        return customersDAO.getAllCustomers();
     }
 
     /**
@@ -221,24 +243,33 @@ public class AdminFacade extends ClientFacade {
      * @param id int
      * @return Customer
      */
-    public Customer getSingleCustomer(int id) {
-        Customer customer = customersDAO.getSingleCustomer(id);
-        List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(id);
-        List<Coupon> customerCoupon = new ArrayList<>();
-        for (CustomersVsCoupons purchase : purchases) {
-            customerCoupon.add(couponsDAO.getSingleCoupon(purchase.getCouponID()));
+    public Customer getSingleCustomer(int id) throws invalidAdminException {
+        if (customersDAO.getSingleCustomer(id) == null) {
+            throw new invalidAdminException("There is no customer with the id: " + id);
         }
-        if (customerCoupon.size() == 0) {
-            return customer;
-        } else {
+        Customer customer = customersDAO.getSingleCustomer(id);
+        List<Coupon> customerCoupon = new ArrayList<>();
+        if (couponsDAO.getAllCustomersCoupons(id) != null) {
+            List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(id);
+            for (CustomersVsCoupons purchase : purchases) {
+                customerCoupon.add(couponsDAO.getSingleCoupon(purchase.getCouponID()));
+            }
+            if (customerCoupon.size() == 0) {
+                return customer;
+            }
             customer.setCoupons(customerCoupon);
         }
         return customer;
     }
 
     public List<CustomersVsCoupons> getAllCustomersVsCoupons(int customerID) throws invalidAdminException {
-        List<CustomersVsCoupons> purchase = customersVsCouponsDAO.getAllCustomersCoupons(customerID);
-        if (purchase.size() == 0) {
+        if (customersDAO.getSingleCustomer(customerID) == null) {
+            throw new invalidAdminException("There is no customer with the id: " + customerID);
+        }
+        List<CustomersVsCoupons> purchase = null;
+        if (couponsDAO.getAllCustomersCoupons(customerID) != null) {
+            purchase = couponsDAO.getAllCustomersCoupons(customerID);
+        } else {
             throw new invalidAdminException("There are no customer purchases!");
         }
         return purchase;
